backend/src/requirements.txt
anthropic==0.40.0
docx==0.2.4
fastapi==0.115.6
langchain==0.3.11
langchain_community==0.3.11
langchain_neo4j==0.1.1
langchain_openai==0.2.12
neo4j==5.27.0
pdfplumber==0.11.4
pydantic==2.9.2
pypandoc==1.14
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.32.1
python-multipart==0.0.19
beautifulsoup4